requirements.txt
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
exceptiongroup==1.1.1
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
outcome==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.7
SQLAlchemy-Utils==0.40.0
tornado==6.2
trio==0.22.0
trio-websocket==0.10.1
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
